+++ packages/ingest-mongodb-public/src/sources/index.ts
@@ -1,6 +1,5 @@
 import { strict as assert } from "assert";
 import { Page, extractFrontMatter } from "mongodb-rag-core";
-import puppeteer from "puppeteer";
 import {
   DataSource,
   makeGitDataSource,
@@ -36,7 +35,10 @@ import {
   prepareWebSources,
   rawWebSources,
 } from "./mongodbDotCom/webSources";
-import { makeWebDataSource } from "./mongodbDotCom/WebDataSource";
+import {
+  makePlaywright,
+  makeWebDataSource,
+} from "./mongodbDotCom/WebDataSource";
 import { chromium } from "playwright";
 
 /**
@@ -186,38 +188,6 @@ const webDataSourceConstructor = async (): Promise<DataSource[]> => {
     sitemapUrls,
   });
 
-  const makePuppeteer = async () => {
-    console.log("hit makePuppeteer:");
-
-    // const executablePath = chromium.executablePath();
-    // console.log('>>>> executablePath:', executablePath);
-
-    const browser = await puppeteer.launch({
-      args: ["--no-sandbox"],
-      headless: "new",
-      // executablePath: executablePath
-      // executablePath: "/opt/homebrew/bin/chromium"
-      // executablePath: "/usr/bin/chromium-browser"
-    });
-    console.log("browser", browser);
-    const page = await browser.newPage();
-    console.log("page", page);
-    return { page, browser };
-  };
-
-  const makePlaywright = async () => {
-    const browserPath = chromium.executablePath();
-    console.log(`Using Chromium executable at: ${browserPath}`);
-    const browser = await chromium.launch({
-      headless: true,
-      executablePath: browserPath,
-    });
-    console.log("playwright browser", browser);
-    const page = await browser.newPage();
-    console.log("playwright page", page);
-    return { page, browser };
-  };
-
   return await Promise.all(
     webSources.map(async (webSource) => {
       return await makeWebDataSource({
@@ -232,6 +202,7 @@ const webDataSourceConstructor = async (): Promise<DataSource[]> => {
   The constructors for the sources used by the docs chatbot.
  */
 export const sourceConstructors: SourceConstructor[] = [
+  webDataSourceConstructor,
   () => makeSnootyDataSources(snootyDataApiBaseUrl, snootyProjectConfig),
   () => makeDevCenterDataSource(devCenterProjectConfig),
   mongoDbUniversitySourceConstructor,
@@ -242,5 +213,4 @@ export const sourceConstructors: SourceConstructor[] = [
   practicalAggregationsDataSource,
   terraformProviderSourceConstructor,
   wiredTigerSourceConstructor,
-  webDataSourceConstructor,
 ];
