+++ packages/ingest-mongodb-public/src/sources/mongodbDotCom/WebDataSource.test.ts
@@ -1,4 +1,3 @@
-import puppeteer, * as Puppeteer from "puppeteer";
 import { makeWebDataSource } from "./WebDataSource";
 import {
   getUrlsFromSitemap,
@@ -79,7 +78,12 @@ describe("prepareWebSources", () => {
   });
 });
 
-test.only("Playwright scraper extracts correct data", async () => {
+describe.skip("makeWebDataSource", () => {
+  // TODO: add tests for this...
+});
+
+// Skipping this test because it requires a browser to be installed
+test.skip("Scraper extracts correct data", async () => {
   const executablePath = chromium.executablePath();
   console.log(">>>> executablePath:", executablePath);
   const browser = await chromium.launch({
